+++ src/FaceMan.DynamicAPI/DynamicWebAPI/ApplicationServiceConvention.cs
@@ -1,5 +1,6 @@
 ﻿using FaceMan.DynamicWebAPI.Config;
 using FaceMan.DynamicWebAPI.Extensions;
+using FaceMan.PublicLibrary;
 
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.ActionConstraints;
