+++ src/FaceMan.DynamicAPI/FaceMan.DynamicAPI.csproj
@@ -1,34 +1,45 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 	<Import Project="..\..\version.props" />
-  <PropertyGroup>
-	  <LangVersion>$(LangVersion)</LangVersion>
-	  <TargetFramework>$(DotNetVersion)</TargetFramework>
-	  <RootNamespace>FaceMan</RootNamespace>
-	  <Title>动态API</Title>
-	  <Authors>FaceMan</Authors>
-	  <PackageProjectUrl>https://github.com/faceman0814/FaceMan.Common/tree/main/src/FaceMan.DynamicAPI</PackageProjectUrl>
-	  <RepositoryUrl>https://github.com/faceman0814/FaceMan.Common</RepositoryUrl>
-	  <Description>Swagger自动扫描Service类生成API文档</Description>
-	  <PackageReadmeFile>readme.md</PackageReadmeFile>
-	  <GenerateDocumentationFile>True</GenerateDocumentationFile>
-	  <Version>1.0.5</Version>
-  </PropertyGroup>
+	<PropertyGroup>
+		<LangVersion>$(LangVersion)</LangVersion>
+		<TargetFramework>$(DotNetVersion)</TargetFramework>
+		<RootNamespace>FaceMan</RootNamespace>
+		<Title>动态API</Title>
+		<Authors>FaceMan</Authors>
+		<PackageProjectUrl>https://github.com/faceman0814/FaceMan.Common/tree/main/src/FaceMan.DynamicAPI</PackageProjectUrl>
+		<RepositoryUrl>https://github.com/faceman0814/FaceMan.Common</RepositoryUrl>
+		<Description>Swagger自动扫描Service类生成API文档</Description>
+		<PackageReadmeFile>readme.md</PackageReadmeFile>
+		<GenerateDocumentationFile>True</GenerateDocumentationFile>
+		<Version>1.0.5</Version>
+	</PropertyGroup>
 
-  <ItemGroup>
-    <!--<PackageReference Include="Microsoft.AspNet.Mvc" Version="5.3.0" />-->
-    <!--<PackageReference Include="Microsoft.AspNetCore.Mvc.Core" Version="2.2.5" />-->
+	<ItemGroup>
+
     <PackageReference Include="Microsoft.AspNetCore.Mvc.Razor.RuntimeCompilation" Version="9.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="9.0.0" />
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
-    <PackageReference Include="Swashbuckle.AspNetCore.Filters" Version="8.0.2" />
-    <PackageReference Include="Swashbuckle.AspNetCore.SwaggerUI" Version="7.0.0" />
-  </ItemGroup>
+		<PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="9.0.0" />
+		<PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
+		<PackageReference Include="Swashbuckle.AspNetCore.Filters" Version="8.0.2" />
+		<PackageReference Include="Swashbuckle.AspNetCore.SwaggerUI" Version="7.0.0" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <None Update="readme.md">
-      <Pack>True</Pack>
-      <PackagePath>\</PackagePath>
-    </None>
-  </ItemGroup>
+	<ItemGroup>
+
+	</ItemGroup>
+	<!-- Common -->
+	<ItemGroup Condition="'$(PublicLibraryRefMode)'=='nuget'">
+		<PackageReference Include="FaceMan.PublicLibrary" Version="$(PublicLibraryVersion)" />
+	</ItemGroup>
+
+	<!-- Common -->
+	<ItemGroup Condition="'$(PublicLibraryRefMode)'=='src'">
+		<ProjectReference Include="..\FaceMan.PublicLibrary\FaceMan.PublicLibrary.csproj" />
+	</ItemGroup>
+	<ItemGroup>
+		<None Update="readme.md">
+			<Pack>True</Pack>
+			<PackagePath>\</PackagePath>
+		</None>
+	</ItemGroup>
 
 </Project>
