+++ src/FaceMan.EntityFrameworkCore/FaceMan.EntityFrameworkCore.csproj
@@ -1,10 +1,10 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net9.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
+	<Import Project="..\..\version.props" />
+	<PropertyGroup>
+		<TargetFramework>$(DotNetVersion)</TargetFramework>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<Nullable>enable</Nullable>
+	</PropertyGroup>
 	<ItemGroup>
 		<PackageReference Include="Microsoft.EntityFrameworkCore" Version="9.0.0" />
 		<PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="9.0.0">
@@ -26,5 +26,15 @@
 		<!--<PackageReference Include="Oracle.EntityFrameworkCore" Version="8.23.60" />-->
 		<!--<PackageReference Include="Pomelo.EntityFrameworkCore.MySql" Version="8.0.2" />-->
 		<PackageReference Include="Microsoft.EntityFrameworkCore.Proxies" Version="9.0.0" />
+	</ItemGroup>
+
+	<!-- Common -->
+	<ItemGroup Condition="'$(PublicLibraryRefMode)'=='nuget'">
+		<PackageReference Include="FaceMan.PublicLibrary" Version="$(PublicLibraryVersion)" />
+	</ItemGroup>
+
+	<!-- Common -->
+	<ItemGroup Condition="'$(PublicLibraryRefMode)'=='src'">
+		<ProjectReference Include="..\FaceMan.PublicLibrary\FaceMan.PublicLibrary.csproj" />
 	</ItemGroup>
 </Project>
